chore: add review comments for socials controller and helpers refactor (pagination, schema counts)
